Complete redcap usernames based on csv-filled-lists with names and email addresses

Complete redcap usernames based on csv-filled-lists with names and email addresses

Usage

redcap_fill_username(which_file = "latest", username_from_mail = FALSE)

Arguments

which_file

as standard REDCap user name list filled out with names and email addresses

username_from_mail

defaults to FALSE. Usernames should be build from First, Last Names. For backward-compatibility.

Value

a csv-file with suffix _completed.csv

Details

Admins have the rights to add additional users to REDCap. The PI or data managers of a project is responsible to fill out the user form and send it via Email to the admin. The admin fills out the usernames according to a standardized rule (handle of the email-adress). When bulk-adding multiple users, this is tedious. This function fills completes the user names automatically and saves the file as a .csv in the same directory as the supplied file.